requirements.txt
Django==3.2
djangorestframework==3.12.2
psycopg2==2.8.6
model-mommy==2.0.0
djangorestframework-simplejwt==4.6.0
